Remove MAINTAINERS file
MAINTAINERS file often quickly falls out-of-date and might be misleading
for those who wish to contribute from outside the core project
developers.

This file was earlier used for e.g. assigning PR code review to
maintainers, but if if the file content is out-of-date, it's not useful
for that purpose. This process has also been guided in CONTRIBUTING.md
so that project owners should be asked for PR review (effectively the
same, but there's no need for specific MAINTAINERS file now).